+++ src/main/java/com/shpakovskiy/soundanalyser/model/Sound.java
@@ -3,9 +3,9 @@ package com.shpakovskiy.soundanalyser.model;
 public class Sound {
     private final float sampleRate;
     private final int sampleSizeBytes; //In some sources "Sample size" is called "Bit depth"
-    private final int[] rawValues;
+    private final double[] rawValues;
 
-    public Sound(float sampleRate, int sampleSizeBytes, int[] rawValues) {
+    public Sound(float sampleRate, int sampleSizeBytes, double[] rawValues) {
         this.sampleRate = sampleRate;
         this.sampleSizeBytes = sampleSizeBytes;
         this.rawValues = rawValues;
@@ -19,7 +19,7 @@ public class Sound {
         return sampleSizeBytes;
     }
 
-    public int[] getRawValues() {
+    public double[] getRawValues() {
         return rawValues;
     }
 }
